+++ list3/src/main/java/com/luq/aisd/list3/Kruskal.java
@@ -15,7 +15,7 @@ public class Kruskal extends MST {
     }
 
     for (int i = 0; i < graph.getV(); i++) {
-      graph.getVerticies().get(i).sort(Comparator.comparingInt(Edge::getWeight));
+      graph.getVerticies().get(i).sort(Comparator.comparingDouble(Edge::getWeight));
     }
 
     for (int i = 0; i < graph.getV(); i++) {
